api/requirements.txt
pip install tensorflow==2.10.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.5.0
matplotlib
numpy